+++ Sources/Models/Extensions/ExtensionType.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public enum ExtensionType {
+    public static let novel = "novel"
+    public static let chineseNovel = "chinese_novel"
+    public static let comic = "comic"
+    public static let tts = "tts"
+}